# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Rapidoo - SaaS B2B de Entregas a Domicilio
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /reports/drivers:
    get:
      operationId: getDriversReport
      tags: [reports]
      summary: Driver performance leaderboard
      parameters:
        - in: query
          name: range
          required: true
          schema:
            type: string
            enum: [day, week, month]
      responses:
        "200":
          description: Driver leaderboard
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DriverPerformance"
...
components:
  schemas:
...
    ZoneName:
      type: string
      enum: [Norte, Sur, Este, Oeste]
...
    DriverPerformance:
      type: object
      properties:
        driverId:
          type: integer
        driverName:
          type: string
        zones:
          type: array
          items:
            $ref: "#/components/schemas/ZoneName"
        deliveries:
          type: integer
        revenue:
          type: number
        active:
          type: boolean
      required: [driverId, driverName, zones, deliveries, revenue, active]
...